% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographiclib-package.R
\docType{package}
\name{geographiclib-package}
\alias{geographiclib}
\alias{geographiclib-package}
\title{geographiclib: Access to 'GeographicLib'}
\description{
GeographicLib functions, starting with MGRS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/geographiclib}
  \item Report bugs at \url{https://github.com/hypertidy/geographiclib/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

Other contributors:
\itemize{
  \item Charles Karney (GeographicLib) [contributor]
}

}
\keyword{internal}
