% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcc.R
\name{lcc0}
\alias{lcc0}
\title{LCC transformation (not useful yet)}
\usage{
lcc0()
}
\value{
numeric area
}
\description{
LCC transformation (not useful yet)
}
\examples{
(code <- mgrs_fwd(c(147.325, -42.881)))
mgrs_rev(code)
}
