% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonarea.R
\name{polygon_area}
\alias{polygon_area}
\title{Polygon area (not useful yet)}
\usage{
polygon_area()
}
\value{
numeric area
}
\description{
Polygon area (not useful yet)
}
\examples{
(code <- mgrs_fwd(c(147.325, -42.881)))
mgrs_rev(code)
}
