% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographiclib-package.R
\docType{package}
\name{geographiclib-package}
\alias{geographiclib}
\alias{geographiclib-package}
\title{geographiclib: Access to 'GeographicLib'}
\description{
Bindings to the 'GeographicLib' C++ library \url{https://geographiclib.sourceforge.io/} for precise geodetic calculations including geodesic computations (distance, bearing, paths, intersections), map projections (UTM/UPS, Transverse Mercator, Lambert Conformal Conic, and more), grid reference systems (MGRS, Geohash, GARS, Georef), coordinate conversions (geocentric, local Cartesian), and polygon area on the WGS84 ellipsoid. All functions are fully vectorized.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/geographiclib}
  \item Report bugs at \url{https://github.com/hypertidy/geographiclib/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

Authors:
\itemize{
  \item Charles Karney (GeographicLib author and maintainer) [copyright holder]
}

Other contributors:
\itemize{
  \item Mark Borgerding (kissfft library (BSD-3-Clause)) [copyright holder]
}

}
\keyword{internal}
