% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographiclib-package.R
\docType{package}
\name{geographiclib-package}
\alias{geographiclib}
\alias{geographiclib-package}
\title{geographiclib: Access to 'GeographicLib'}
\description{
GeographicLib functions for precise geodetic calculations including MGRS (Military Grid Reference System), UTM/UPS projections, Geohash encoding, geodesic computations (distance, bearing, paths), and polygon area on the WGS84 ellipsoid. All functions are fully vectorized with rich metadata output.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/geographiclib}
  \item Report bugs at \url{https://github.com/hypertidy/geographiclib/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

Authors:
\itemize{
  \item Charles Karney (GeographicLib author and maintainer) [copyright holder]
}

}
\keyword{internal}
