% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgrs.R
\name{mgrs_fwd}
\alias{mgrs_fwd}
\alias{mgrs_rev}
\title{MGRS conversion}
\usage{
mgrs_fwd(x, precision = 5L)

mgrs_rev(code)
}
\arguments{
\item{x}{in forward mode a matrix of lon,lat coordinates}

\item{precision}{in forward mode integer between 0 and 5 (default is 5, full precision)}

\item{code}{in reverse mode an MGRS code string}
}
\value{
lon,lat,x,y,zone,northp vector in reverse mode, MGRS code in forward mode
}
\description{
Forward mode is vectorized on \code{x} by rows and \code{precision}.
}
\details{
Currently cannnot handle missing values.
}
\examples{
(code <- mgrs_fwd(cbind(147.325, -42.881)))
mgrs_rev(code)
x <- cbind(long = c(-63.22, 34.02, 49.45, 45.67, 47.4, -66.3, -65.72, 42.31, 8.94, 102.55),
           lat = c(17.62, -1.9, 37.47, 39.84, 33.15, 17.98, -21.44, 41.65, 50.08, 24.36))
codes <- mgrs_fwd(x, precision = rep(c(0, 5), each = 5))
}
