% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgrs.R
\name{mgrs_fwd}
\alias{mgrs_fwd}
\alias{mgrs_rev}
\title{MGRS}
\usage{
mgrs_fwd(x)

mgrs_rev(code)
}
\arguments{
\item{x}{in forward mode a lon,lat coordinate}

\item{code}{in reverse mode an MGRS code string}
}
\value{
lon,lat,x,y,zone,northp vector in reverse mode, MGRS code in forward mode
}
\description{
MGRS
}
\examples{
(code <- mgrs_fwd(c(147.325, -42.881)))
mgrs_rev(code)
}
